player: don't request speed, playtime, position and repeatShuffle on creation
Don't request the speed, playtime, position and repeatShuffle properties of the
player when it gets created. These will automatically be requested when the
player gets active, as refresh() will be called then and it will also request
those properties. Also fixes a lot of error messages when starting the app.