lib: work around Player.OnStop timing issues
When the Player.OnStop notification gets send for the music player an
immediate Player.GetActivePlayers() call would still return the audio
player as being active. This means we'd have to wait a bit before making
the call. 500ms is on the (very) safe side as it already works with 50ms
using the emulator on the same machine as XBMC and running on a WIFI
connected/mobile device the timeout could probably be set even lower due
to added network latency.