sailfish: add 'Wake host' button to no connection page
Currently it's cumbersome (and not intuive) to wake a host after
starting the app, as we don't want to host to be woken without user
intervention. This fixes that by adding a 'Wake host' button on the
default 'Connection' page. So the user can wake the host without having
to navigate through some menus.