simplify playItem implementation
By moving the implementation to playing 'something' to the
player the calling code gets much simpler as it doesn't
need to care about playlists etc anymore.

Default implementation uses Player.Open so we can add support
for resume. AudioPlayer::open still uses the playlist based
implementation as that's more convenient to the user.