update ubuntu build system
Now that the Ubuntu SDK also supports qmake cross compiling, update
the build system to do that. Also dropping the old
run_on_ubuntu_touch script that has been used to build packages
and run on the phone so far. As Ubuntu now supports click packages
on all systems (not just phones) it doesn't look like the
debian packaging is going to be used any more, so dropping that too.