conn: fix check if timed out command is version request
m_commandId is the id to use for the next command, so it would never
match with any command, as there is no command with that id yet. Besides
it would still be faulty if there were other commands created (send in
parallel, or over socket, or queued) in the meantime.