sailfish: fix controls dock showing after reconnect
When you're connected to a Kodi host which is playing something, and then reconnect
to another Kodi host, which isn't playing something, the controls dock would be shown
inadverently. This fixes that by checking if there still is an active player when
a dialog closes and the dock would get reopened.