videos: refresh when an item in the list is updated
Fixes issue were the resume state wouldn't be updated when playback
of an episode/movie/musicvideo was stopped. Which is needed to
correctly show the marker (as for unplayed) and to ask the user to
resume playback when he wants to play it again.

As 'resume' isn't part of the notification we'd have to fetch the ful
item again, and refreshing the full list makes things a lot easier than
updating just the item.